requirements.txt
python-dotenv~=1.1.0
python-telegram-bot==22.0
SQLAlchemy~=2.0.40
APScheduler~=3.11.0